Split piston handlers in subfolder package.
